Break up code into small files. Massive refactoring
