+++ src/interface/paths.jl
@@ -1,0 +1,5 @@
+module paths
+
+    const ROOT_DIR::String = abspath(dirname(abspath(@__FILE__)), "..", "..")
+
+end
