+++ test/get_coverage.jl
@@ -1,6 +1,6 @@
 #!/usr/bin/env -S julia --color=yes --startup-file=no
 # Run this function from inside the AGNI root folder
-# e.g. as `julia --project=. tests/get_coverage.jl `
+# e.g. as `julia --project=. test/get_coverage.jl`
 
 ROOT_DIR = abspath(joinpath(dirname(abspath(@__FILE__)),"../"))
 
@@ -12,6 +12,7 @@ Pkg.activate(ROOT_DIR)
 # https://github.com/JuliaCI/Coverage.jl
 
 # Include libraries
+using LoggingExtras
 using Coverage
 using Printf
 using Dates
@@ -194,11 +195,11 @@ open(output_file, "w") do io
     println(io, "_Generated by `test/get_coverage.jl`_")
 end
 
-println("Coverage report written to: $output_file")
-println(@sprintf("Overall coverage: %.1f%% (%d / %d lines)",
-    overall_pct, total_covered, total_lines))
+@info "Coverage report written to: $output_file"
+@info @sprintf("Overall coverage: %.1f%% (%d / %d lines)",
+                overall_pct, total_covered, total_lines)
 
-println("Uploading to Codecov...")
+@info "Uploading to Codecov..."
 Coverage.upload_to_codecov(coverage)
 
-println("Done")
+@info "Done"
